+++ apps/docs/src/app/documentation/core-helpers/toolbar/toolbar.component.ts
@@ -15,6 +15,7 @@ import {
     FD_LANGUAGE_ALBANIAN,
     FD_LANGUAGE_BULGARIAN,
     FD_LANGUAGE_ENGLISH,
+    FD_LANGUAGE_FRENCH,
     FD_LANGUAGE_GEORGIAN,
     FD_LANGUAGE_ITALIAN,
     FD_LANGUAGE_RUSSIAN,
@@ -71,7 +72,7 @@ export class ToolbarDocsComponent implements OnInit, OnDestroy {
         // { name: 'Hrvatski', value: FD_LANGUAGE_CROATIAN }, TODO: uncomment when translations are provided
         // { name: 'Český', value: FD_LANGUAGE_CZECH }, TODO: uncomment when translations are provided
         { name: 'English', value: FD_LANGUAGE_ENGLISH },
-        // { name: 'Français', value: FD_LANGUAGE_FRENCH }, TODO: uncomment when translations are provided
+        { name: 'Français', value: FD_LANGUAGE_FRENCH },
         { name: 'ქართული', value: FD_LANGUAGE_GEORGIAN },
         // { name: 'Deutsch', value: FD_LANGUAGE_GERMAN }, TODO: uncomment when translations are provided
         // { name: 'עִברִית', value: FD_LANGUAGE_HEBREW }, TODO: uncomment when translations are provided