+++ apps/docs/src/app/platform/component-docs/platform-table/platform-table-examples/platform-table-filterable-example.component.ts
@@ -121,8 +121,6 @@ export class TableDataProviderExample extends TableDataProvider<ExampleItem> {
     private filter({ filterBy }: TableState): ExampleItem[] {
         let items = this.items;
 
-        console.log(filterBy);
-
         filterBy
             .filter(({ field }) => !!field)
             .forEach((rule) => {