+++ libs/platform/src/lib/dynamic-page/dynamic-page.module.ts
@@ -14,24 +14,33 @@ import { DynamicPageLayoutActionsComponent } from './dynamic-page-header/actions
 import { DynamicPageHeaderComponent } from './dynamic-page-header/header/dynamic-page-header.component';
 import { DynamicPageKeyInfoComponent } from './dynamic-page-header/key-info/dynamic-page-key-info.component';
 import { DynamicPageTitleComponent } from './dynamic-page-header/title/dynamic-page-title.component';
-import { DynamicPageTitleHostComponent } from './dynamic-page-header/title/dynamic-page-title-host.component';
+import { DynamicPageTitleImageComponent } from './dynamic-page-header/title/dynamic-page-title-image.component';
 import { DynamicPageComponent } from './dynamic-page.component';
 import { ScrollingModule } from '@angular/cdk/scrolling';
+import { FacetModule } from '@fundamental-ngx/core/facets';
 
 @NgModule({
     declarations: [
         DynamicPageComponent,
         DynamicPageTitleComponent,
-        DynamicPageTitleHostComponent,
         DynamicPageKeyInfoComponent,
         DynamicPageGlobalActionsComponent,
         DynamicPageLayoutActionsComponent,
         DynamicPageHeaderComponent,
         DynamicPageContentComponent,
         DynamicPageContentHostComponent,
-        DynamicPageFooterComponent
+        DynamicPageFooterComponent,
+        DynamicPageTitleImageComponent
     ],
-    imports: [CommonModule, TabsModule, DynamicPageModule, PlatformButtonModule, BreadcrumbModule, ScrollingModule],
+    imports: [
+        CommonModule,
+        TabsModule,
+        DynamicPageModule,
+        PlatformButtonModule,
+        BreadcrumbModule,
+        ScrollingModule,
+        FacetModule
+    ],
     exports: [
         DynamicPageComponent,
         DynamicPageTitleComponent,
@@ -40,7 +49,8 @@ import { ScrollingModule } from '@angular/cdk/scrolling';
         DynamicPageLayoutActionsComponent,
         DynamicPageHeaderComponent,
         DynamicPageContentComponent,
-        DynamicPageFooterComponent
+        DynamicPageFooterComponent,
+        DynamicPageTitleImageComponent
     ]
 })
 export class PlatformDynamicPageModule {}