+++ CHANGELOG.md
@@ -1,4 +1,29 @@
 
+## [0.36.0-rc.17](https://github.com/SAP/fundamental-ngx/compare/v0.36.0-rc.16...v0.36.0-rc.17) (2022-08-15)
+
+
+### Features
+
+* **core,platform:** bring French texts for i18n ([#8510](https://github.com/SAP/fundamental-ngx/issues/8510)) ([a595580](https://github.com/SAP/fundamental-ngx/commit/a5955805f6adcfb22248929aabafe300990ae180))
+
+## [0.36.0-rc.16](https://github.com/SAP/fundamental-ngx/compare/v0.36.0-rc.15...v0.36.0-rc.16) (2022-08-15)
+
+
+### ⚠ BREAKING CHANGES
+
+* **platform:** * fdp-dynamic-page-title-host component removed as not used
+
+### Bug Fixes
+
+* **platform:** dynamic page facets and title content, refactoring ([#8527](https://github.com/SAP/fundamental-ngx/issues/8527)) ([2e93e29](https://github.com/SAP/fundamental-ngx/commit/2e93e295f57c3643931a9f8ad5c32d29e06157cb))
+
+## [0.36.0-rc.15](https://github.com/SAP/fundamental-ngx/compare/v0.36.0-rc.14...v0.36.0-rc.15) (2022-08-15)
+
+
+### Features
+
+* **platform:** responsive table columns ([#8532](https://github.com/SAP/fundamental-ngx/issues/8532)) ([d9ecfe4](https://github.com/SAP/fundamental-ngx/commit/d9ecfe4845f5724f9099afbd7ae0c8b1ba472be9))
+
 ## [0.36.0-rc.14](https://github.com/SAP/fundamental-ngx/compare/v0.36.0-rc.13...v0.36.0-rc.14) (2022-08-11)